add field data exists route, move functions to services
